pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esm-agentbench"
version = "0.1.0"
description = "Certified AI Reasoning - Spectral Certificates for LLM agents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ESM AgentBench Contributors", email = "andreuballus@gmail.com"}
]
keywords = ["llm", "agent-safety", "spectral-analysis", "certification", "koopman-operator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# pyproject.toml is canonical; requirements.txt is for convenience installs.
dependencies = [
    "huggingface-hub==0.16.4",
    "sentence-transformers==2.3.1",
    "flask==3.0.0",
    "numpy==1.26.2",
    "scipy==1.11.4",
    "scikit-learn==1.3.2",
    "joblib==1.3.2",
    "pytest==7.4.3",
    "tomli==2.0.1",
    "gunicorn==20.1.0",
    "matplotlib==3.8.1",
    "seaborn==0.13.0",
    "pandas==2.2.3",
    "pytest-asyncio==0.21.1",
    "PyYAML==6.0.1",
    "pydantic==2.5.0",
    "openai==1.3.0",
    "jupyter==1.0.0",
    "nbconvert==7.11.0",
    "transformers==4.35.2",
    "accelerate==0.25.0",
    "torch>=2.2.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
quantization = [
    "bitsandbytes==0.41.3",
]
inference = [
    "vllm==0.2.7",
]
signing = [
    "PyNaCl==1.5.0",
]

[project.urls]
Homepage = "https://github.com/ipsissima/esm-agentbench"
Repository = "https://github.com/ipsissima/esm-agentbench"
Issues = "https://github.com/ipsissima/esm-agentbench/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["assessor*", "certificates*", "esmassessor*", "scenarios*", "benchmarks*", "tools*"]
exclude = ["tests*", "analysis*", "demo_swe*", "legacy*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.md"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
